Enzo src-directory reorganization: following some of James's suggestions part 3 -> moved some files into ``gravity/solvers`` and ``gravity/matrix``
I opted to create ``gravity/solvers`` and ``gravity/matrix`` rather than ``numerical/solvers`` and ``numerical/matrix`` since the gravity solver is currently the only thing that makes use of this multigrid functionality (but if we ever add other physics-solvers that use multigrid functionality, I think those other names are an excellent idea)
