+++ src/Enzo/gravity/CMakeLists.txt
@@ -11,6 +11,8 @@
 # - See the CMake Primer section of the developer documentation for more details
 file(GLOB LOCAL_SRC_FILES CONFIGURE_DEPENDS
   *.cpp *.hpp *.F
+  matrix/*.cpp matrix/*.hpp
+  solvers/*.cpp solvers/*.hpp
 )
 
 target_sources(enzo PRIVATE ${LOCAL_SRC_FILES})
